+++ guillotina/interfaces/misc.py
@@ -40,10 +40,12 @@ class IAbsoluteURL(Interface):
 
 
 class IAddOn(Interface):
+    @classmethod
     def install(cls, container, request):  # noqa: N805
         """
         """
 
+    @classmethod
     def uninstall(cls, container, request):  # noqa: N805
         """
         """