+++ guillotina/request.py
@@ -32,6 +32,8 @@ class Request(web_request.Request):
     exc = None
     view_name = None
     found_view = None
+    resource = None
+    tail = None
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)