+++ guillotina/contrib/cache/strategy.py
@@ -75,8 +75,6 @@ class BasicCache(BaseCache):
         await self._utility.delete_all(keys)
 
     async def store_object(self, obj, pickled):
-        if not self.push_enabled:
-            return
         if len(self._stored_objects) < self.max_publish_objects:
             self._stored_objects.append((obj, pickled))
             # also assume these objects are then stored
@@ -92,7 +90,7 @@ class BasicCache(BaseCache):
         return invalidated
 
     @profilable
-    async def close(self, invalidate=True):
+    async def close(self, invalidate=True, publish=True):
         if self._utility is None:
             return
         try:
@@ -107,9 +105,9 @@ class BasicCache(BaseCache):
                 )
                 await self.delete_all(keys_to_invalidate)
 
-            if len(self._keys_to_publish) > 0 and self._utility._subscriber is not None:
-                keys = self._keys_to_publish
-                asyncio.ensure_future(self.synchronize(keys))
+                if publish and len(self._keys_to_publish) > 0 and self._utility._subscriber is not None:
+                    keys = self._keys_to_publish
+                    asyncio.ensure_future(self.synchronize(keys))
             self._keys_to_publish = []
         except Exception:
             logger.warning("Error closing connection", exc_info=True)
@@ -133,9 +131,10 @@ class BasicCache(BaseCache):
                 ob_key = self.get_key(container=obj.__parent__, id=obj.__name__)
                 await self.set(val, container=obj.__parent__, id=obj.__name__)
 
-            if ob_key in keys_to_publish:
-                keys_to_publish.remove(ob_key)
-            push[ob_key] = val
+            if self.push_enabled:
+                if ob_key in keys_to_publish:
+                    keys_to_publish.remove(ob_key)
+                push[ob_key] = val
 
         self._utility.ignore_tid(self._transaction._tid)
         await self._utility._subscriber.publish(