+++ guillotina/interfaces/response.py
@@ -7,7 +7,7 @@ class IResponse(Interface):
     content = Attribute("content")
     headers = Attribute("headers")
 
-    def __init__(content=None, headers=None):
+    def __init__(*, content: dict = None, headers: dict = None, status: int = None):
         """
         """